Convert recent announcements to new format
Include all releases and news entries up to the CI migration notice from
August 2023.  This unifies the release notes format a bit and also acts
as a PoC for the new format.

We may convert the other entries in the future.  For now, keep the rest
in a separate archive.md file and just include it on the timeline page.

While at it, tweak some of the overview sections (mostly in prereleases)
so that they're more consistent.